Fix scroll position
When you share a link to a pattern or anti-pattern, the scroll position places some of the content beneath the sticky-header. For example: https://web.devopstopologies.com/#anti-type-f

This also removes the need to introduce smooth scrolling via JavaScript as the `#` links will smooth scroll where the user has no preference for reduced motion, though I haven't submitted any changes to JavaScript. I can take a look if that's wanted.